api & test refactoring
